Python package updates for command language

+++ tesseract/tesseract_collision/include/tesseract_collision/test_suite/collision_box_cone_unit.hpp
@@ -149,12 +149,12 @@ inline void runTest(DiscreteContactManager& checker)
   }
   else
   {
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][0], 0.5, 0.001);
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][1], 0.0, 0.001);
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][2], -0.125, 0.001);
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][0], -0.05, 0.001);
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][1], 0.0, 0.001);
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][2], -0.125, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][0], 0.5, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][1], 0.0, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][2], -0.125, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][0], -0.05, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][1], 0.0, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][2], -0.125, 0.001);
   }
 
   EXPECT_NEAR(result_vector[0].normal[0], idx[2] * 1.0, 0.001);
@@ -207,12 +207,12 @@ inline void runTest(DiscreteContactManager& checker)
   }
   else
   {
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][0], 0.5, 0.001);
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][1], 0.0, 0.001);
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][2], -0.125, 0.001);
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][0], 0.75, 0.001);
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][1], 0.0, 0.001);
-    EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][2], -0.125, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][0], 0.5, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][1], 0.0, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][2], -0.125, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][0], 0.75, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][1], 0.0, 0.001);
+    EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][2], -0.125, 0.001);
   }
 
   EXPECT_NEAR(result_vector[0].normal[0], idx[2] * 1.0, 0.001);
